feat: implement simulation service and request DTOs for sorting, searching, and pathfinding algorithms

+++ backend/src/main/java/com/algorithmrace/visualizer/dto/LaneStats.java
@@ -1,4 +1,4 @@
 package com.algorithmrace.visualizer.dto;
 
 public record LaneStats(
-    int comparisons, int swaps, int steps, long timeMs, boolean found, Integer foundIndex) {}
+    long comparisons, long swaps, long steps, long timeMs, boolean found, Integer foundIndex) {}
